changed doubles to ints

+++ src/SportProperty.java
@@ -3,13 +3,13 @@ public class SportProperty extends Property
     // properties
 
     // constructors
-    public SportProperty( double price, double[] rents )
+    public SportProperty( int price, int[] rents )
     {
         super(price,rents);
     }
 
     // methods
-    public double calculateRent()
+    public int calculateRent()
     {
         int sportPropertyCount = getOwner().getSportPropertyCount();
         return getRents()[sportPropertyCount - 1];
